Move parse/tojson methods to subclasses
