+++ crud/src/main/java/com/redhat/lightblue/crud/BulkRequest.java
@@ -18,11 +18,18 @@
  */
 package com.redhat.lightblue.crud;
 
+import java.util.List;
+import java.util.ArrayList;
+import java.util.Iterator;
+import java.util.Collections;
+import java.util.Comparator;
+
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.fasterxml.jackson.databind.node.ArrayNode;
 import com.fasterxml.jackson.databind.node.JsonNodeFactory;
 
+import com.redhat.lightblue.util.JsonObject;
 import com.redhat.lightblue.Request;
 
 /**
@@ -42,25 +49,17 @@ import com.redhat.lightblue.Request;
  */
 public class BulkRequest extends AbstractBulkJsonObject<Request> {
 
-    @Override
     public JsonNode toJson() {
-        return toJson("requests","request");
+        throw new UnsupportedOperationException();
     }
-
+    
     public static BulkRequest fromJson(ObjectNode node) {
         BulkRequest req=new BulkRequest();
         req.parse((ArrayNode)node.get("requests"));
         return req;
     }
     
-    @Override
-    protected void toJsonEntryNode(ObjectNode node,Request entry) {
-        node.set("op",JsonNodeFactory.instance.textNode(entry.getOperation().toString()));
-        node.set("request",entry.toJson());
-    }
 
-
-    @Override
     protected Request parseEntry(ObjectNode node) {
         JsonNode opNode=node.get("op");
         if(opNode!=null) {
@@ -86,4 +85,53 @@ public class BulkRequest extends AbstractBulkJsonObject<Request> {
         } else
             throw new IllegalArgumentException("op");
     }    
+
+    private static class SReq {
+        private final int seq;
+        private final Request entry;
+
+        public SReq(int seq,Request entry) {
+            this.seq=seq;
+            this.entry=entry;
+        }
+    }
+
+    /**
+     * Parses the bulk object from the given array node
+     */
+    protected void parse(ArrayNode entriesArray) {
+        entries.clear();
+        // Fill the entries into an array list, assuming for most
+        // cases the array list will be ordered by the sequence
+        // number. If there are out-of-sequence entries, then sort it
+        List<SReq> list=new ArrayList<>(entriesArray.size());
+        int lastSeq=0;
+        boolean first=true;
+        boolean ooo=false;
+        for(Iterator<JsonNode> itr=entriesArray.elements();itr.hasNext();) {
+            JsonNode x=itr.next();
+            JsonNode val=x.get("seq");
+            if(val!=null) {
+                int seq=val.asInt();
+                if(first)
+                    lastSeq=seq;
+                else {
+                    if(seq<lastSeq)
+                        ooo=true;
+                    else
+                        lastSeq=seq;
+                }
+                Request entry=parseEntry((ObjectNode)x);
+                list.add(new SReq(seq,entry));
+            }
+        }
+        if(ooo)
+            Collections.sort(list,new Comparator<SReq>() {
+                    public int compare(SReq r1,SReq r2) {
+                        return r1.seq-r2.seq;
+                    }
+                });
+        for(SReq r:list)
+            entries.add(r.entry);
+    }
 }
